Error fix
- Fixed an error that would happen due to not having done a sign in a long time

- The user has to reauthenticate to be able to delete his account

- If the user chooses a different account the account won't be deleted and a snackbar message will be shown
